+++ phoenix/monitor/views/__init__.py
@@ -1,9 +1,28 @@
+import requests
+
+from pyramid.view import view_config
 from pyramid.events import subscriber
+from pyramid.response import Response
 
 from phoenix.events import JobFinished, JobStarted
 
 import logging
-logger = logging.getLogger(__name__)
+LOGGER = logging.getLogger(__name__)
+
+
+@view_config(route_name='wpsoutputs')
+def wpsoutputs(request):
+    filename = request.matchdict.get('filename')
+    url = request.registry.settings.get('wps.output.url')
+    url += '/wpsoutputs'
+    url += filename
+    LOGGER.debug("delegate to wpsoutputs: %s", url)
+    # forward request to target (without Host Header)
+    # h = dict(request.headers)
+    # h.pop("Host", h)
+    resp = requests.get(url, verify=False)
+    resp.raise_for_status()
+    return Response(resp.content, status=resp.status_code)
 
 
 @subscriber(JobStarted)
@@ -15,9 +34,9 @@ def notify_job_started(event):
 @subscriber(JobFinished)
 def notify_job_finished(event):
     if event.succeeded():
-        logger.info("job %s succeded.", event.job.get('title'))
+        LOGGER.info("job %s succeded.", event.job.get('title'))
         # event.request.session.flash("Job <b>{0}</b> succeded.".format(event.job.get('title')), queue='success')
     else:
-        logger.warn("job %s failed.", event.job.get('title'))
+        LOGGER.warn("job %s failed.", event.job.get('title'))
         # logger.warn("status = %s", event.job.get('status'))
         # event.request.session.flash("Job <b>{0}</b> failed.".format(event.job.get('title')), queue='danger')